requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
scipy>=1.10.0
h5py>=3.8.0
matplotlib>=3.7.0
tqdm>=4.65.0
pyyaml>=6.0

# Utilities
joblib>=1.2.0

# Development
pytest>=7.3.0
black>=23.3.0
flake8>=6.0.0

# Optional: Experiment tracking
# wandb>=0.15.0
# tensorboard>=2.13.0